BAU: Add timestamp to versions
Concourse will only trigger if a resource's check returns a version it
hasn't seen before.

This meant that if the GitHub team membership changed to a combination
of members previously encountered (including no members at all) this
resources wouldn't trigger.

To resolve this problem, this change appends a timestamp to the version
returned to Concourse. This means that each version is unique and the
resource will trigger on every membership change.